+++ src/Arborescence.Traversal/EnumerableBfs/EnumerableBfs.EnumerateVertices.MultipleSource.cs
@@ -1,9 +1,10 @@
 namespace Arborescence.Traversal
 {
+    using System;
     using System.Collections.Generic;
 
     public readonly partial struct EnumerableBfs<
-        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TGraphPolicy, TExploredSetPolicy>
+        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TExploredSetPolicy>
     {
         /// <summary>
         /// Enumerates vertices of the graph in a breadth-first order starting from the multiple sources.
@@ -13,10 +14,20 @@ namespace Arborescence.Traversal
         /// <param name="exploredSet">The set of explored vertices.</param>
         /// <typeparam name="TVertexEnumerator">The type of the vertex enumerator.</typeparam>
         /// <returns>An enumerator to enumerate the vertices of the the graph.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="graph"/> is <see langword="null"/>,
+        /// or <paramref name="sources"/> is <see langword="null"/>.
+        /// </exception>
         public IEnumerator<TVertex> EnumerateVertices<TVertexEnumerator>(
             TGraph graph, TVertexEnumerator sources, TExploredSet exploredSet)
             where TVertexEnumerator : IEnumerator<TVertex>
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
+            if (sources == null)
+                throw new ArgumentNullException(nameof(sources));
+
             var queue = new Internal.Queue<TVertex>();
             try
             {
@@ -30,11 +41,11 @@ namespace Arborescence.Traversal
 
                 while (queue.TryTake(out TVertex u))
                 {
-                    TEdgeEnumerator outEdges = GraphPolicy.EnumerateOutEdges(graph, u);
+                    TEdgeEnumerator outEdges = graph.EnumerateOutEdges(u);
                     while (outEdges.MoveNext())
                     {
                         TEdge e = outEdges.Current;
-                        if (!GraphPolicy.TryGetHead(graph, e, out TVertex v))
+                        if (!graph.TryGetHead(e, out TVertex v))
                             continue;
 
                         if (ExploredSetPolicy.Contains(exploredSet, v))