+++ src/Arborescence.Traversal/EnumerableDfs/ReverseDfs.EnumerateVertices.SingleSource.cs
@@ -1,9 +1,10 @@
 namespace Arborescence.Traversal
 {
+    using System;
     using System.Collections.Generic;
 
     public readonly partial struct ReverseDfs<
-        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TGraphPolicy, TExploredSetPolicy>
+        TGraph, TVertex, TEdge, TEdgeEnumerator, TExploredSet, TExploredSetPolicy>
     {
         /// <summary>
         /// Enumerates vertices of the graph in a depth-first order starting from the single source.
@@ -12,8 +13,14 @@ namespace Arborescence.Traversal
         /// <param name="source">The source.</param>
         /// <param name="exploredSet">The set of explored vertices.</param>
         /// <returns>An enumerator to enumerate the vertices of the the graph.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="graph"/> is <see langword="null"/>.
+        /// </exception>
         public IEnumerator<TVertex> EnumerateVertices(TGraph graph, TVertex source, TExploredSet exploredSet)
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
             var stack = new Internal.Stack<TVertex>();
             try
             {
@@ -27,11 +34,11 @@ namespace Arborescence.Traversal
                     ExploredSetPolicy.Add(exploredSet, u);
                     yield return u;
 
-                    TEdgeEnumerator outEdges = GraphPolicy.EnumerateOutEdges(graph, u);
+                    TEdgeEnumerator outEdges = graph.EnumerateOutEdges(u);
                     while (outEdges.MoveNext())
                     {
                         TEdge e = outEdges.Current;
-                        if (!GraphPolicy.TryGetHead(graph, e, out TVertex v))
+                        if (!graph.TryGetHead(e, out TVertex v))
                             continue;
 
                         stack.Add(v);