+++ src/Arborescence.Traversal/GenericSearch/GenericSearch.EnumerateVertices.MultipleSource.cs
@@ -1,9 +1,10 @@
 ﻿namespace Arborescence.Traversal
 {
+    using System;
     using System.Collections.Generic;
 
     public readonly partial struct GenericSearch<TGraph, TVertex, TEdge, TEdgeEnumerator, TFringe,
-        TExploredSet, TGraphPolicy, TFringePolicy, TExploredSetPolicy>
+        TExploredSet, TFringePolicy, TExploredSetPolicy>
     {
         /// <summary>
         /// Enumerates vertices of the graph in an order specified by the fringe starting from the multiple sources.
@@ -14,10 +15,20 @@
         /// <param name="exploredSet">The set of explored vertices.</param>
         /// <typeparam name="TVertexEnumerator">The type of the vertex enumerator.</typeparam>
         /// <returns>An enumerator to enumerate the vertices of the the graph.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="graph"/> is <see langword="null"/>,
+        /// or <paramref name="sources"/> is <see langword="null"/>.
+        /// </exception>
         public IEnumerator<TVertex> EnumerateVertices<TVertexEnumerator>(
             TGraph graph, TVertexEnumerator sources, TFringe fringe, TExploredSet exploredSet)
             where TVertexEnumerator : IEnumerator<TVertex>
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
+            if (sources == null)
+                throw new ArgumentNullException(nameof(sources));
+
             while (sources.MoveNext())
             {
                 TVertex source = sources.Current;
@@ -32,11 +43,11 @@
                 ExploredSetPolicy.Add(exploredSet, u);
                 yield return u;
 
-                TEdgeEnumerator outEdges = GraphPolicy.EnumerateOutEdges(graph, u);
+                TEdgeEnumerator outEdges = graph.EnumerateOutEdges(u);
                 while (outEdges.MoveNext())
                 {
                     TEdge e = outEdges.Current;
-                    if (!GraphPolicy.TryGetHead(graph, e, out TVertex v))
+                    if (!graph.TryGetHead(e, out TVertex v))
                         continue;
 
                     FringePolicy.Add(fringe, v);