+++ src/Arborescence.Traversal/GenericSearch/GenericSearch.EnumerateVertices.SingleSource.cs
@@ -1,9 +1,10 @@
 ﻿namespace Arborescence.Traversal
 {
+    using System;
     using System.Collections.Generic;
 
     public readonly partial struct GenericSearch<TGraph, TVertex, TEdge, TEdgeEnumerator, TFringe,
-        TExploredSet, TGraphPolicy, TFringePolicy, TExploredSetPolicy>
+        TExploredSet, TFringePolicy, TExploredSetPolicy>
     {
         /// <summary>
         /// Enumerates vertices of the graph in an order specified by the fringe starting from the single source.
@@ -13,9 +14,15 @@
         /// <param name="fringe">The collection of discovered vertices which are not finished yet.</param>
         /// <param name="exploredSet">The set of explored vertices.</param>
         /// <returns>An enumerator to enumerate the vertices of the the graph.</returns>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="graph"/> is <see langword="null"/>.
+        /// </exception>
         public IEnumerator<TVertex> EnumerateVertices(
             TGraph graph, TVertex source, TFringe fringe, TExploredSet exploredSet)
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
             FringePolicy.Add(fringe, source);
 
             while (FringePolicy.TryTake(fringe, out TVertex u))
@@ -26,11 +33,11 @@
                 ExploredSetPolicy.Add(exploredSet, u);
                 yield return u;
 
-                TEdgeEnumerator outEdges = GraphPolicy.EnumerateOutEdges(graph, u);
+                TEdgeEnumerator outEdges = graph.EnumerateOutEdges(u);
                 while (outEdges.MoveNext())
                 {
                     TEdge e = outEdges.Current;
-                    if (!GraphPolicy.TryGetHead(graph, e, out TVertex v))
+                    if (!graph.TryGetHead(e, out TVertex v))
                         continue;
 
                     FringePolicy.Add(fringe, v);