+++ src/Arborescence.Traversal/InstantDfs/RecursiveDfs.MultipleSource.cs
@@ -3,8 +3,7 @@
     using System;
     using System.Collections.Generic;
 
-    public readonly partial struct RecursiveDfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TGraphPolicy,
-        TColorMapPolicy>
+    public readonly partial struct RecursiveDfs<TGraph, TVertex, TEdge, TEdgeEnumerator, TColorMap, TColorMapPolicy>
     {
         /// <summary>
         /// Traverses the graph in a DFS manner starting from the multiple sources.
@@ -19,7 +18,8 @@
         /// <typeparam name="TVertexEnumerator">The type of the vertex enumerator.</typeparam>
         /// <typeparam name="THandler">The type of the events handler.</typeparam>
         /// <exception cref="ArgumentNullException">
-        /// <paramref name="sources"/> is <see langword="null"/>,
+        /// <paramref name="graph"/> is <see langword="null"/>,
+        /// or <paramref name="sources"/> is <see langword="null"/>,
         /// or <paramref name="handler"/> is <see langword="null"/>.
         /// </exception>
         public void Traverse<TVertexEnumerator, THandler>(
@@ -27,6 +27,9 @@
             where TVertexEnumerator : IEnumerator<TVertex>
             where THandler : IDfsHandler<TGraph, TVertex, TEdge>
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
             if (sources == null)
                 throw new ArgumentNullException(nameof(sources));
 
@@ -59,7 +62,8 @@
         /// <typeparam name="TVertexEnumerator">The type of the vertex enumerator.</typeparam>
         /// <typeparam name="THandler">The type of the events handler.</typeparam>
         /// <exception cref="ArgumentNullException">
-        /// <paramref name="sources"/> is <see langword="null"/>,
+        /// <paramref name="graph"/> is <see langword="null"/>,
+        /// or <paramref name="sources"/> is <see langword="null"/>,
         /// or <paramref name="handler"/> is <see langword="null"/>.
         /// </exception>
         public void Traverse<TVertexEnumerator, THandler>(
@@ -67,6 +71,9 @@
             where TVertexEnumerator : IEnumerator<TVertex>
             where THandler : IDfsHandler<TGraph, TVertex, TEdge>
         {
+            if (graph == null)
+                throw new ArgumentNullException(nameof(graph));
+
             if (sources == null)
                 throw new ArgumentNullException(nameof(sources));